+++ functions/ListOutput.fnc.php
@@ -459,7 +459,9 @@ function ListOutput( $result, $column_names, $singular = '.', $plural = '.', $li
 					button(
 						'add',
 						$link['add']['title'],
-						'"' . URLEscape( $link['add']['link'] ) . '"'
+						( mb_strpos( $link['add']['link'], '"' ) === 0 ?
+							$link['add']['link'] :
+							'"' . URLEscape( $link['add']['link'] ) . '"' )
 					) . '</td></tr>';
 			}
 			elseif ( $link['add']['span'] && ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
@@ -598,7 +600,9 @@ function ListOutput( $result, $column_names, $singular = '.', $plural = '.', $li
 				button(
 					'add',
 					issetVal( $link['add']['title'], '' ),
-					$link['add']['link']
+					( mb_strpos( $link['add']['link'], '"' ) === 0 ?
+						$link['add']['link'] :
+						'"' . URLEscape( $link['add']['link'] ) . '"' )
 				) . '</td></tr>';
 			}
 			elseif ( isset( $link['add']['span'] ) && ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
@@ -647,7 +651,9 @@ function ListOutput( $result, $column_names, $singular = '.', $plural = '.', $li
 			button(
 				'add',
 				issetVal( $link['add']['title'], '' ),
-				$link['add']['link']
+				( mb_strpos( $link['add']['link'], '"' ) === 0 ?
+					$link['add']['link'] :
+					'"' . URLEscape( $link['add']['link'] ) . '"' )
 			) . '</div>';
 		}
 		elseif (  ( ! empty( $link['add']['html'] )